requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[postgresql]
python-dotenv
pydantic-settings
asyncpg
psycopg2